using a common Synchronizer class on top of label/milestone classes:
a few stuff was copy and paste
so I figured they could belong to an AbstractSynchronizer class
that the LabelSynchronizer and MilestoneSynchronizer extend
